# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GYM API Documentation",
        "description": "GYM API Documentation for make connections to your application",
        "contact": {
            "name": "Brenbala",
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/units-of-measurement": {
            "get": {
                "tags": [
                    "Units of Measurement"
                ],
                "summary": "Get all units of measurement",
                "description": "Returns a list of all units of measurement.",
                "operationId": "getUnitsOfMeasurement",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/user-types": {
            "get": {
                "tags": [
                    "User Types"
                ],
                "summary": "Get all user types",
                "description": "Returns a list of all user types.",
                "operationId": "getUserTypes",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
...
}